Fix more Coverity issues in the PHY lib
CID 369560 Missing unlock, in dft_fftw.c
CID 363710 Logically dead code, in resampler.c
CID 55097 Result is not floating point, in chest_test_dl.c
CID 351020 Result is not floating point, in chest_nbiot_test_dl.c

+++ lib/src/phy/ch_estimation/test/chest_test_dl.c
@@ -137,7 +137,7 @@ int main(int argc, char** argv)
       for (uint32_t n_port = 0; n_port < cell.nof_ports; n_port++) {
         srsran_vec_cf_zero(input, num_re);
         for (i = 0; i < num_re; i++) {
-          input[i] = 0.5 - rand() / RAND_MAX + I * (0.5 - rand() / RAND_MAX);
+          input[i] = 0.5 - rand() / (float)RAND_MAX + I * (0.5 - rand() / (float)RAND_MAX);
         }
 
         srsran_vec_cf_zero(ce, num_re);
